requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
db-dtypes==1.0.4
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.7.1
google-cloud-bigquery==3.3.6
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandas-gbq==0.17.9
pandera==0.13.4
proto-plus==1.22.1
protobuf==4.21.9
psycopg2==2.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
